requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
asyncpg==0.30.0
attrs==25.3.0
et_xmlfile==2.0.0
frozenlist==1.6.0
greenlet==3.2.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
propcache==0.3.1
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.20.0
